Restructure Tooltip
