+++ projects/js-packages/charts/src/components/tooltip/private/base-tooltip.module.scss
